requirements.txt
flask_sqlalchemy
flask
python-dotenv